feat(gui): add component organization into pages and page-controllers.
Pages:
A Page is a single screen in the application. It only needs to specify its Scene, that is its graphical representation inside a stage.
In particular, an FXMLPage is a Page whose graphical representation is determined by an fxml file, while the logic behind the user
interactions is handled by an FXMLController. So, an FXMLPage only requires an fxml file and an FXMLController in input.
Note: fxml is not yet supported by scalafx, therefore certain procedures have been delegated to JavaFX.

A Page can be extended with several mixins, for now only:
- ApplicablePage: it makes it possible to change the page of the application by simply applying the page constructor to the Stage.
- PageWithController: it adds a Controller to the page.

Examples of the applications of these traits are in the three now-available main screens of the graphical user interface:
MainMenuPage, GameConfigurationPage and GamePage.

Controllers:
A Controller handles the logic of a graphical component and must have access to the Stage in order to apply changes to the
graphical user interface.
In particular, an FXMLController is also Initializable and exports the @FXML tag by default. The @FXML tag is used by
ScalaFX/JavaFX for injecting the references to certain graphical components into the FXMLController. Due to the nature of
the injection procedure, these components must be declared as "var" and must be initialized as "null"/"_" (causing alerts from
wartremover).

Examples of these traits are in the three now-available controllers for the afore-mentioned pages: MainMenuController,
GameConfigurationController, GamePageController.
Note: fields annotated with @FXML must have the same name of a component in the corresponding fxml file. Their action listener
      must be declared in the Initializable.initialize method, which is guaranteed to be called after the injection procedure.

More:
Page and Controller are both FXComponent. An FXComponent includes most of the implicits of ScalaFX: these are needed to
interoperate with JavaFX, so that, for example, a ScalaFX.Button is interchangable with a JavaFX.Button.
In particular, a StageComponent is a FXComponent that has also access to the stage of the application (i.e. Controller,
ApplicablePage...).

The entry point of the application is ChessGameInterface, whose method launch(...) starts the graphical user interface.
Through launch(...) it would be possible (after minimal refactoring) to pass some inputs to the graphical user interface
(such as the chess engine service).
In the stageConfiguration(...), the primary stage is configured (i.e. title, size and more todo).

Debug:
A util.debug.Logger has been added for logging to the console with contextual and more detailed information.
A util.debug.Exception.observe has been added for observing a block of code and printing its first exception.